catalog/ui/scripts/requirements-mockup.txt
# Optional: only for regenerating docs/images/ops-showroom-panel-mockup.png
pillow>=10.0.0
